code review fixes - rename ts_split and ts_concat files
