+++ src/plugins/intel_gna/src/transformations/ts_concat_forward.hpp
